# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API для Todo и Auth",
    "version": "1.0.0",
    "description": "Документация API для управления задачами и пользователями"
  },
  "servers": [
    {
      "url": "https://expresslesson1-last.onrender.com/api/v1",
      "description": "Продакшн сервер"
    },
    {
      "url": "http://localhost:5000/api/v1",
      "description": "local сервер"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Регистрация пользователя",
        "operationId": "registerUser",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "name": {
                    "type": "string"
                  },
                  "profilePhoto": {
                    "type": "string",
                    "format": "uri",
                    "description": "Ссылка на фото профиля (необязательно)"
                  }
                },
                "required": ["email", "password", "name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь успешно зарегистрирован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Пользователь успешно зарегистрирован"
                    },
                    "accessToken": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации или некорректные данные",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Некорректный email или пароль"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Пользователь с таким email уже существует",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Этот email уже используется"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/getProfile": {
      "get": {
        "summary": "Получить профиль пользователя",
        "operationId": "getProfile",
        "tags": ["Auth"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Профиль пользователя успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "password": {
                          "type": "string"
                        },
                        "profilePhoto": {
                          "type": "string"
                        },
                        "isAdmin": {
                          "type": "boolean"
                        },
                        "products": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "title": {
                                "type": "string"
                              },
                              "price": {
                                "type": "number"
                              },
                              "description": {
                                "type": "string"
                              },
                              "image": {
                                "type": "string"
                              },
                              "authorEmail": {
                                "type": "string"
                              },
                              "isCompleted": {
                                "type": "boolean"
                              },
                              "createdAt": {
                                "type": "string"
                              },
                              "updatedAt": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/updateProfile": {
      "put": {
        "summary": "Обновить профиль пользователя",
        "operationId": "updateProfile",
        "tags": ["User"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "doni@gmail.com"
                  },
                  "profilePhoto": {
                    "type": "string",
                    "example": "https://uznayvse.ru/images/content/2023/7/24/singer-doni_88.jpg"
                  },
                  "name": {
                    "type": "string",
                    "example": "doni"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Профиль успешно обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Profile updated successfully"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 15
                        },
                        "name": {
                          "type": "string",
                          "example": "balancha"
                        },
                        "email": {
                          "type": "string",
                          "example": "Balancha@gmail.com"
                        },
                        "profilePhoto": {
                          "type": "string",
                          "example": "https://uznayvse.ru/images/content/2023/7/24/singer-doni_88.jpg"
                        },
                        "isAdmin": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}